Removed old readonly files, rename new
