+++ src/views/ModelPlan/ReadOnly/Payments/index.tsx
@@ -11,9 +11,9 @@ import { ModelInfoContext } from 'views/ModelInfoWrapper';
 import { NotFoundPartial } from 'views/NotFound';
 
 import ReadOnlyBody from '../_components/Body';
-import ReadOnlySectionNew, {
+import ReadOnlySection, {
   RelatedUnneededQuestions
-} from '../_components/ReadOnlySection/new';
+} from '../_components/ReadOnlySection';
 import { getRelatedUneededQuestions } from '../_components/ReadOnlySection/util';
 import TitleAndStatus from '../_components/TitleAndStatus';
 import { ReadOnlyProps } from '../ModelBasics';
