+++ src/views/ModelPlan/ReadOnly/_components/ReadOnlySection/index.test.tsx
@@ -1,66 +1,84 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import { FundingSource } from 'gql/gen/graphql';
+import { FundingSource, OverlapType, RecruitmentType } from 'gql/gen/graphql';
 import i18next from 'i18next';
 
+import { participantsAndProviders } from 'i18n/en-US/modelPlan/participantsAndProviders';
 import { payments } from 'i18n/en-US/modelPlan/payments';
+import { Bool } from 'types/translation';
 
-import ReadOnlySection, {
+import ReadOnlySection from '.';
+import {
+  compareClosure,
   formatListItems,
   formatListOtherItems,
-  formatListTooltips
-} from './index';
+  formatListTooltips,
+  getRelatedUneededQuestions,
+  isHiddenByParentCondition
+} from './util';
+
+const defaultProps = {
+  field: 'modelApplicationLevel',
+  translations: participantsAndProviders,
+  values: {
+    providerOverlap: OverlapType.NO,
+    providerOverlapHierarchy: 'Overlap heirarchy',
+    modelApplicationLevel: 'Top level',
+    recruitmentMethod: [RecruitmentType.OTHER],
+    recruitmentOther: 'Other recruitment'
+  }
+};
 
 describe('The Read Only Section', () => {
   describe('As a Non-list Component', () => {
-    const defaultCopyProps = {
-      heading: 'React Testing is Great',
-      copy: 'Lorem ipsum dolor sit amet.'
-    };
-
     it('renders without crashing', async () => {
-      render(<ReadOnlySection {...defaultCopyProps} />);
+      render(<ReadOnlySection {...defaultProps} />);
 
-      expect(screen.getByText(defaultCopyProps.heading)).toBeInTheDocument();
-      expect(screen.getByText(defaultCopyProps.copy)).toBeInTheDocument();
+      expect(
+        screen.getByText(
+          i18next.t<string>(
+            'participantsAndProviders:modelApplicationLevel.label'
+          )
+        )
+      ).toBeInTheDocument();
+      expect(screen.getByText('Top level')).toBeInTheDocument();
     });
 
     it('renders "No additional information specified" if copy is empty', async () => {
-      render(<ReadOnlySection {...defaultCopyProps} copy={null} />);
+      const emptyData = { ...defaultProps };
 
-      expect(screen.getByText(defaultCopyProps.heading)).toBeInTheDocument();
+      emptyData.values.modelApplicationLevel = '';
+
+      render(<ReadOnlySection {...emptyData} />);
+
       expect(
-        screen.getByText('No additional information specified')
+        screen.getByText(
+          i18next.t<string>(
+            'participantsAndProviders:modelApplicationLevel.label'
+          )
+        )
+      ).toBeInTheDocument();
+      expect(
+        screen.getByText(
+          i18next.t<string>('miscellaneous:noAdditionalInformation')
+        )
       ).toBeInTheDocument();
     });
   });
 
   describe('As a List Component', () => {
-    const defaultListProps = {
-      heading: 'Lorem ipsum dolor sit amet',
-      list: true,
-      listItems: ['Center for Medicare (CM)', 'CMMI']
-    };
-
-    it('renders without crashing', async () => {
-      render(<ReadOnlySection {...defaultListProps} />);
-
-      expect(screen.getByText(defaultListProps.heading)).toBeInTheDocument();
-      expect(
-        screen.getByText(defaultListProps.listItems[0])
-      ).toBeInTheDocument();
-    });
-
     it('renders Other entry', async () => {
-      render(
-        <ReadOnlySection
-          {...defaultListProps}
-          listItems={['Other']}
-          copy="Lorem ipsum dolor sit amet."
-        />
-      );
+      const listData = { ...defaultProps };
 
-      expect(screen.getByText(defaultListProps.heading)).toBeInTheDocument();
+      listData.field = 'recruitmentMethod';
+
+      render(<ReadOnlySection {...listData} />);
+
+      expect(
+        screen.getByText(
+          i18next.t<string>('participantsAndProviders:recruitmentMethod.label')
+        )
+      ).toBeInTheDocument();
       expect(screen.getByText('Other')).toBeInTheDocument();
       expect(screen.getByTestId('other-entry')).toBeInTheDocument();
     });
@@ -130,6 +148,57 @@ describe('The Read Only Section', () => {
       expect(formatListTooltips(payments.fundingSource, values)).toEqual(
         expectedOrder
       );
+    });
+
+    it('gets a list of realted, unneeded questions', async () => {
+      const value: OverlapType[] = [OverlapType.NO];
+
+      const expectedQuestions: string[] = [
+        i18next.t<string>(
+          'participantsAndProviders:providerOverlapHierarchy.label'
+        )
+      ];
+
+      const relatedQuestions = getRelatedUneededQuestions(
+        participantsAndProviders.providerOverlap,
+        value
+      );
+
+      expect(relatedQuestions).toEqual(expectedQuestions);
+    });
+
+    it('return true for question that is hidden and conditional on a parent question', async () => {
+      const expectedEvaluation = true;
+
+      const childQuestion = isHiddenByParentCondition(
+        participantsAndProviders.providerOverlapHierarchy,
+        defaultProps.values
+      );
+
+      expect(childQuestion).toEqual(expectedEvaluation);
+    });
+
+    it('return false for question is shown and not conditional', async () => {
+      const expectedEvaluation = false;
+
+      const childQuestion = isHiddenByParentCondition(
+        participantsAndProviders.providerOverlap,
+        defaultProps.values
+      );
+
+      expect(childQuestion).toEqual(expectedEvaluation);
+    });
+
+    it('compares closures of parent/child translation configuration, child should now be hidden', async () => {
+      const expectedEvaluation = false;
+
+      const childQuestion = compareClosure(
+        Bool.true,
+        participantsAndProviders.gainsharePayments,
+        participantsAndProviders.gainsharePaymentsTrack
+      );
+
+      expect(childQuestion).toEqual(expectedEvaluation);
     });
   });
 });
